backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pydantic
pydantic-settings
anthropic
python-docx
python-dotenv
